pyproject.toml
[tool.poetry]
name = "jager-test"
version = "0.1.0"
description = ""
authors = ["Danis Sharafiev <danis.sharafiev@bk.ru>"]
readme = "README.md"
packages = [{include = "jager_test"}]

[tool.poetry.dependencies]
python = "3.8.6"
fastapi = "^0.99.1"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
opentelemetry-sdk = "^1.18.0"
opentelemetry-exporter-jaeger = "^1.18.0"
opentelemetry-instrumentation-fastapi = "^0.39b0"
opentelemetry-instrumentation-aiohttp-client = "^0.39b0"
aiohttp = "^3.8.4"
opentelemetry-exporter-otlp = "^1.18.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.18.0"
aiokafka = "^0.8.1"
asyncpg = "^0.27.0"
aioredis = "^2.0.1"
punq = "^0.6.2"
backoff = "^2.2.1"
loguru = "^0.7.0"
aiocache = {extras = ["redis"], version = "^0.12.1"}
opentelemetry-instrumentation-redis = "^0.39b0"
email-validator = "^2.0.0.post2"
alembic = "^1.11.1"
opentelemetry-instrumentation-sqlalchemy = "^0.39b0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
